se actualizan las paradas de estaciones de metro y se concatenan archivos para generar uno con datos de un día desagregado por hora.

+++ codigo/procesar_datos.sh
@@ -18,6 +18,8 @@ GENERAR_VIAJE_CSV=false
 GENERAR_VIAJE_CON_ETAPAS_CSV=false
 # crea los archivos pajek a partir de los archivos csv
 GENERAR_PAJEK=false
+# crea el csv de paradas con toda su información (nombre, longitud, latitud, ...)
+GENERAR_CSV_PARADAS=false
 # genera las comunidades en dos niveles, por medio del framework infomap
 GENERAR_COMUNIDADES=false
 # concatena los archivos creados por hora en un solo archivo para mostrar una secuencia en el cartodb
@@ -78,9 +80,7 @@ chmod 775 -R $RUTA_DATOS
   # Estos deben tener la siguiente sintaxis XX-YY [XX-YY ...]
   # donde XX e YY son números enteros de dos dígitos en el rango [00-23]
   # y pueden ser iguales. Para concatenar varios tramos se usa el espacio. Ej: XX-YY ZZ-TT
-  TRAMOS=(01-01 02-02 03-03 04-04 05-05 06-06 07-07 08-08 09-09 10-10 11-11 12-12 13-13 14-14 15-15 16-16 17-17 18-18 19-19 20-20 21-21 22-22 23-23 00-00)
-  # Tramos transantiago
-  #TRAMOS=(XX-XX XX-XX XX-XX)
+  TRAMOS=(01-01 02-02 03-03 04-04 05-05 06-06 07-07 08-08 09-09 10-10 11-11 12-12 13-13 14-14 15-15 16-16 17-17 18-18 19-19 20-20 21-21 22-22 23-23 00-00 06-09 18-21)
 
   # para filtrar por hora usar        : extract(hour from tiempo_subida)
   # para filtrar por fecha y hora usar: (date_trunc('hour', tiempo_subida))
@@ -434,17 +434,21 @@ if [ "$GENERAR_PAJEK" = true ]; then
   done
 fi
 
-# Genera un csv con los datos de la tabla parada_util. Es usado al final del proceso para actualizar
-# los códigos de paraderos e insertar su respectiva posición geográfica.
-NOMBRE_PARADAS_CSV="PARADAS.csv"
-PARADAS_CSV="copy (SELECT * FROM parada_util) To '$RUTA_DATOS/$NOMBRE_PARADAS_CSV' WITH DELIMITER ';' CSV;"
-rm -f $RUTA_DATOS/$NOMBRE_PARADAS_CSV
-sudo -u postgres -i psql -d memoria -c "$PARADAS_CSV"
+
+if [ "$GENERAR_CSV_PARADAS" = true ]; then
+
+  # Genera un csv con los datos de la tabla parada_util. Es usado al final del proceso para actualizar
+  # los códigos de paraderos e insertar su respectiva posición geográfica.
+  NOMBRE_PARADAS_CSV="PARADAS.csv"
+  PARADAS_CSV="copy (SELECT * FROM parada_util) To '$RUTA_DATOS/$NOMBRE_PARADAS_CSV' WITH DELIMITER ';' CSV;"
+  rm -f $RUTA_DATOS/$NOMBRE_PARADAS_CSV
+  sudo -u postgres -i psql -d memoria -c "$PARADAS_CSV"
+fi 
 
 if [ "$GENERAR_COMUNIDADES" = true ]; then
   ####################################################################################
   rm -f $RUTA_DATOS_INFOMAP/*.tree
-  rm -f $RUTA_DATOS_CARTODB/*.csv
+  rm -f -R $RUTA_DATOS_CARTODB/*.csv
 
   for ARCHIVO_NET in $RUTA_DATOS_PAJEK/*.net; do
     echo "Procesando $ARCHIVO_NET"
@@ -455,30 +459,46 @@ if [ "$GENERAR_COMUNIDADES" = true ]; then
     HORA=$(echo "$NOMBRE_INFOMAP" | cut -d '-' -f 1)
     if [[ "$NOMBRE_INFOMAP" == "*semana*" ]]; then
       FECHA="SEMANA"
+      CARPETA="semana_etapa"
     elif [[ $NOMBRE_INFOMAP == *lunes_a_jueves* ]]; then
       FECHA="2013-04-17T$HORA:00:00Z"
+      CARPETA="lunes_a_jueves_etapa"
     elif [[ "$NOMBRE_INFOMAP" == *viernes* ]]; then
       FECHA="2013-04-18T$HORA:00:00Z"
+      CARPETA="viernes_etapa"
     elif [[ "$NOMBRE_INFOMAP" == *sabado* ]]; then
       FECHA="2013-04-19T$HORA:00:00Z"
+      CARPETA="sabado_etapa"
     elif [[ $NOMBRE_INFOMAP == *domingo* ]]; then
       FECHA="2013-04-20T$HORA:00:00Z"
+      CARPETA="domingo_etapa"
     fi
 
     # -i 'pajek'  Indica el formato de archivo de entrada
     # --two-level Optimiza una partición de dos niveles de la red
     # -d          Asume que los arcos tienen dirección
     $RUTA_INFOMAP/Infomap -i 'pajek' --two-level -d "$ARCHIVO_NET" "$RUTA_DATOS_INFOMAP"
-    php tree2csv.php $RUTA_DATOS/$NOMBRE_PARADAS_CSV $RUTA_DATOS_INFOMAP/$NOMBRE_INFOMAP.tree $RUTA_DATOS_CARTODB $FECHA
+    php tree2csv.php $RUTA_DATOS/$NOMBRE_PARADAS_CSV $RUTA_DATOS_INFOMAP/$NOMBRE_INFOMAP.tree $RUTA_DATOS_CARTODB/$CARPETA $FECHA
   done 
 fi
 
 if [ "$CONCATENAR_HORAS" = true ]; then
   ####################################################################################
-  cat $RUTA_DATOS_CARTODB/*domingo*.csv >> $RUTA_DATOS_CARTODB/domingo.csv
-  cat $RUTA_DATOS_CARTODB/*sabado*.csv >> $RUTA_DATOS_CARTODB/sabado.csv
-  cat $RUTA_DATOS_CARTODB/*lunes_a_jueves*.csv >> $RUTA_DATOS_CARTODB/lunes_a_jueves.csv
-  cat $RUTA_DATOS_CARTODB/*semana*.csv >> $RUTA_DATOS_CARTODB/semana.csv
+  rm -f $RUTA_DATOS_CARTODB/lunes_a_jueves.csv
+
+  ENCABEZADO="1er_nivel 2do_nivel PageRank Nombre latitud longitud fecha id_pajek"
+  ARCHIVOS=("lunes_a_jueves" "viernes" "sabado" "domingo")
+
+  for ARCHIVO in ${ARCHIVOS[@]}; do
+    echo "Procesando datos $ARCHIVO ..."
+    cat $RUTA_DATOS_CARTODB/${ARCHIVO}_etapa/*.csv >> $RUTA_DATOS_CARTODB/$ARCHIVO.csv
+    # se eliminan los encabezados del archivo(uno por archivo concatenado)
+    sed -i '/^1er_nivel/d' $RUTA_DATOS_CARTODB/$ARCHIVO.csv
+    # se agrega encabezado
+    echo "$ENCABEZADO" | cat - $RUTA_DATOS_CARTODB/$ARCHIVO.csv > $RUTA_DATOS_CARTODB/tmp && mv $RUTA_DATOS_CARTODB/tmp $RUTA_DATOS_CARTODB/$ARCHIVO.csv
+  done
+  
+  rm -f tmp
 fi 
 
 # cambiamos el dueño de los archivos para poder verlos en el entorno de escritorio
